refactor: split observation port DTOs (#62)
