+++ backend/internal/ports/doc.go
@@ -1,0 +1,5 @@
+// Package ports declares boundary interfaces and DTOs used to connect core
+// services to replaceable adapters such as runtimes, workspaces, trackers, and
+// storage writers. Domain models stay in internal/domain; generated storage rows
+// stay inside storage packages.
+package ports
